IP-1097 adding some documentation
