+++ docs/models.html
@@ -107,6 +107,7 @@
 </li>
 </ul>
 </li>
+<li class="toctree-l2"><a class="reference internal" href="index.html#installation">Installation</a></li>
 <li class="toctree-l2"><a class="reference internal" href="index.html#getting-started">Getting started</a></li>
 </ul>
 </li>
