+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["index","models","validation"],envversion:52,filenames:["index.rst","models.rst","validation.rst"],objects:{},objnames:{},objtypes:{},terms:{"4dabae77118d":0,"case":[],"class":0,"default":0,"import":0,"super":0,"true":0,"while":0,The:0,Then:0,Use:0,__init__:0,_blank:[],_metadata:1,_version:0,abov:0,access:0,actual:0,affect:0,against:0,agnost:0,aim:0,align:0,all:0,allow:0,alq:0,also:0,alternatebas:0,analysisreturnuri:0,ani:0,annot:0,antrun:0,applic:0,apt:0,archiv:0,arg:0,argument:0,asserttru:[],audit:[],auditlog:1,aug:0,automat:0,avail:0,avoid:0,avpr2html:0,avro:0,avro_factori:0,avrodoc:0,base:0,batch:0,between:0,biodata:0,box:0,build_model:0,can:0,cancer:[],cancerinterpretedgenom:0,cancerparticip:1,catalog_variable_set:0,chang:[],check:0,cip:0,cipapi:0,clinic:0,clinicalreportcanc:1,clinicalreportrd:1,clone:0,code:0,com:0,command:0,comment:1,common:1,commoninterpret:1,commoninterpretedcanc:1,commonparticip:1,commonrequest:1,complet:0,configur:0,containerid:0,convers:0,copi:0,correspond:0,could:0,coverag:0,creat:0,create_batch:0,cva:0,cva_0_3_0:0,data:[],dataintak:[],dataintakecanc:1,dataintakerd:1,def:0,demograph:0,depend:0,dependency_manag:[],deploi:0,describ:0,deserialis:0,differ:0,directori:0,diseas:[],django:0,doc:0,docker:0,drwxr:0,e444d27c16b9:0,each:0,either:0,enabl:0,england:0,ensur:0,errr:0,evid:[0,1],evidenceset:1,exampl:0,exchang:0,exit:0,exitquestionnair:1,exitquestionnairebackup:[],extern:0,facilit:0,factori:0,factoryavro:0,field:0,file:0,fill:0,fill_nul:0,flag:0,folder:0,follow:0,format:0,fuzz:0,fuzzi:0,fuzzychoic:0,fuzzyinteg:0,ga4gh:0,ga4gh_3_0_0:0,ga4ghvariantfactori:0,gel:[],gel_bioinf_model:0,gel_models_tool:0,gelbammetr:1,gelmodelstool:0,gelprotocol:0,gelreportmodel:[0,2],gelvcfmetr:1,gener:0,genericfactoryavro:0,genom:0,genomicsengland:0,get_factory_avro:0,git:0,github:0,greatli:0,handi:0,have:0,healthcheck:1,help:0,hoc:0,hold:0,how:[],href:[],html:0,html_schema:0,http:0,idl2avpr:0,idl2json:0,idl2python:0,idl:0,illustr:0,implement:0,includ:0,independ:0,index:0,individualst:1,inform:0,instal:[],instanc:0,instance_with_uri:0,intak:[],interfac:0,intern:0,interpret:0,interpretation_request_factori:[],interpretationdatard:[],interpretationrequestcanc:1,interpretationrequestrd:[0,1],interpretedgenomescanc:1,interpretedgenomesrd:1,ish:0,isvalid:0,its:0,jar:0,json2java:0,json2python:0,json:0,json_fil:2,kwarg:0,lambda:0,land:0,languag:0,last:[0,2],latest:[0,2],lazyattribut:0,librari:0,list:0,log:[],mai:0,map:0,marker:0,maven:0,mdtdeliveryprotocol:1,messag:0,meta:0,metadata:1,method:1,metric:0,metrics_1_0_0:0,modelvalidatior:2,mvn:0,myuri:0,need:0,nodej:0,non:0,note:0,npm:0,nullabl:0,object:0,observ:[],observedvari:1,onc:0,one:0,onli:0,opencb:0,opencb_1_2_0:0,org:0,org_ga4gh_models_3_0_0_common:[],org_ga4gh_models_3_0_0_metadata:[],org_ga4gh_models_3_0_0_method:[],org_ga4gh_models_3_0_0_read:[],org_ga4gh_models_3_0_0_readmethod:[],org_ga4gh_models_3_0_0_refer:[],org_ga4gh_models_3_0_0_referencemethod:[],org_ga4gh_models_3_0_0_vari:[],org_ga4gh_models_3_0_0_variantmethod:[],org_ga4gh_models_3_1_0:[],org_gel_models_cva_avro_0_3_1_com:[],org_gel_models_cva_avro_0_3_1_dataintak:[],org_gel_models_cva_avro_0_3_1_evidenceset:[],org_gel_models_cva_avro_0_3_1_observedvari:[],org_gel_models_cva_avro_0_3_1_reportedvari:[],org_gel_models_cva_avro_0_4_0:[],org_gel_models_metrics_avro_1_0_0_gelbammetr:[],org_gel_models_metrics_avro_1_0_0_gelvcfmetr:[],org_gel_models_metrics_avro_1_0_0_individualst:[],org_gel_models_metrics_avro_1_0_0_rarediseaseinterpretationpipelin:[],org_gel_models_metrics_avro_1_0_0_samplest:[],org_gel_models_metrics_avro_1_0_0_supplementaryanalysisresult:[],org_gel_models_metrics_avro_1_0_1_gelbammetr:[],org_gel_models_metrics_avro_1_0_1_gelvcfmetr:[],org_gel_models_metrics_avro_1_0_1_individualst:[],org_gel_models_metrics_avro_1_0_1_rarediseaseinterpretationpipelin:[],org_gel_models_metrics_avro_1_0_1_samplest:[],org_gel_models_metrics_avro_1_0_1_supplementaryanalysisresult:[],org_gel_models_metrics_avro_1_1_0:[],org_gel_models_participant_avro_1_0_0_cancerparticip:[],org_gel_models_participant_avro_1_0_0_commonparticip:[],org_gel_models_participant_avro_1_0_0_participantsensitiveinform:[],org_gel_models_participant_avro_1_0_0_rdparticip:[],org_gel_models_participant_avro_1_0_0_rdparticipantchangelog:[],org_gel_models_participant_avro_1_0_0_versioncontrol:[],org_gel_models_participant_avro_1_0_1_cancerparticip:[],org_gel_models_participant_avro_1_0_1_commonparticip:[],org_gel_models_participant_avro_1_0_1_participantsensitiveinform:[],org_gel_models_participant_avro_1_0_1_rdparticip:[],org_gel_models_participant_avro_1_0_1_rdparticipantchangelog:[],org_gel_models_participant_avro_1_0_1_versioncontrol:[],org_gel_models_participant_avro_1_0_2_cancerparticip:[],org_gel_models_participant_avro_1_0_2_commonparticip:[],org_gel_models_participant_avro_1_0_2_participantsensitiveinform:[],org_gel_models_participant_avro_1_0_2_rdparticip:[],org_gel_models_participant_avro_1_0_2_rdparticipantchangelog:[],org_gel_models_participant_avro_1_0_2_versioncontrol:[],org_gel_models_participant_avro_1_0_3:[],org_gel_models_report_avro_2_1_0_auditlog:[],org_gel_models_report_avro_2_1_0_cancerparticip:[],org_gel_models_report_avro_2_1_0_clinicalreportcanc:[],org_gel_models_report_avro_2_1_0_clinicalreportrd:[],org_gel_models_report_avro_2_1_0_commoninterpret:[],org_gel_models_report_avro_2_1_0_commoninterpretedcanc:[],org_gel_models_report_avro_2_1_0_commonparticip:[],org_gel_models_report_avro_2_1_0_commonrequest:[],org_gel_models_report_avro_2_1_0_exitquestionnair:[],org_gel_models_report_avro_2_1_0_gelbammetr:[],org_gel_models_report_avro_2_1_0_gelvcfmetr:[],org_gel_models_report_avro_2_1_0_individualst:[],org_gel_models_report_avro_2_1_0_interpretationrequestcanc:[],org_gel_models_report_avro_2_1_0_interpretationrequestrd:[],org_gel_models_report_avro_2_1_0_interpretedgenomescanc:[],org_gel_models_report_avro_2_1_0_interpretedgenomesrd:[],org_gel_models_report_avro_2_1_0_mdtdeliveryprotocol:[],org_gel_models_report_avro_2_1_0_participantsensitiveinform:[],org_gel_models_report_avro_2_1_0_rdparticip:[],org_gel_models_report_avro_2_1_0_rdparticipantchangelog:[],org_gel_models_report_avro_2_1_0_samplest:[],org_gel_models_report_avro_2_1_0_supplementaryanalysisresult:[],org_gel_models_report_avro_2_1_0_versioncontrol:[],org_gel_models_report_avro_3_0_0_auditlog:[],org_gel_models_report_avro_3_0_0_cancerparticip:[],org_gel_models_report_avro_3_0_0_clinicalreportcanc:[],org_gel_models_report_avro_3_0_0_clinicalreportrd:[],org_gel_models_report_avro_3_0_0_commoninterpret:[],org_gel_models_report_avro_3_0_0_commoninterpretedcanc:[],org_gel_models_report_avro_3_0_0_commonparticip:[],org_gel_models_report_avro_3_0_0_commonrequest:[],org_gel_models_report_avro_3_0_0_exitquestionnair:[],org_gel_models_report_avro_3_0_0_gelbammetr:[],org_gel_models_report_avro_3_0_0_gelvcfmetr:[],org_gel_models_report_avro_3_0_0_individualst:[],org_gel_models_report_avro_3_0_0_interpretationrequestcanc:[],org_gel_models_report_avro_3_0_0_interpretationrequestrd:[],org_gel_models_report_avro_3_0_0_interpretedgenomescanc:[],org_gel_models_report_avro_3_0_0_interpretedgenomesrd:[],org_gel_models_report_avro_3_0_0_mdtdeliveryprotocol:[],org_gel_models_report_avro_3_0_0_participantsensitiveinform:[],org_gel_models_report_avro_3_0_0_rdparticip:[],org_gel_models_report_avro_3_0_0_rdparticipantchangelog:[],org_gel_models_report_avro_3_0_0_samplest:[],org_gel_models_report_avro_3_0_0_supplementaryanalysisresult:[],org_gel_models_report_avro_3_0_0_versioncontrol:[],org_gel_models_report_avro_3_1_0_auditlog:[],org_gel_models_report_avro_3_1_0_cancerparticip:[],org_gel_models_report_avro_3_1_0_clinicalreportcanc:[],org_gel_models_report_avro_3_1_0_clinicalreportrd:[],org_gel_models_report_avro_3_1_0_commoninterpret:[],org_gel_models_report_avro_3_1_0_commoninterpretedcanc:[],org_gel_models_report_avro_3_1_0_commonparticip:[],org_gel_models_report_avro_3_1_0_commonrequest:[],org_gel_models_report_avro_3_1_0_exitquestionnair:[],org_gel_models_report_avro_3_1_0_gelbammetr:[],org_gel_models_report_avro_3_1_0_gelvcfmetr:[],org_gel_models_report_avro_3_1_0_individualst:[],org_gel_models_report_avro_3_1_0_interpretationrequestcanc:[],org_gel_models_report_avro_3_1_0_interpretationrequestrd:[],org_gel_models_report_avro_3_1_0_interpretedgenomescanc:[],org_gel_models_report_avro_3_1_0_interpretedgenomesrd:[],org_gel_models_report_avro_3_1_0_mdtdeliveryprotocol:[],org_gel_models_report_avro_3_1_0_participantsensitiveinform:[],org_gel_models_report_avro_3_1_0_rdparticip:[],org_gel_models_report_avro_3_1_0_rdparticipantchangelog:[],org_gel_models_report_avro_3_1_0_samplest:[],org_gel_models_report_avro_3_1_0_supplementaryanalysisresult:[],org_gel_models_report_avro_3_1_0_versioncontrol:[],org_gel_models_report_avro_4_0_0_auditlog:[],org_gel_models_report_avro_4_0_0_cancerparticip:[],org_gel_models_report_avro_4_0_0_clinicalreportcanc:[],org_gel_models_report_avro_4_0_0_clinicalreportrd:[],org_gel_models_report_avro_4_0_0_commoninterpret:[],org_gel_models_report_avro_4_0_0_commoninterpretedcanc:[],org_gel_models_report_avro_4_0_0_commonrequest:[],org_gel_models_report_avro_4_0_0_exitquestionnair:[],org_gel_models_report_avro_4_0_0_gelbammetr:[],org_gel_models_report_avro_4_0_0_interpretationrequestcanc:[],org_gel_models_report_avro_4_0_0_interpretationrequestrd:[],org_gel_models_report_avro_4_0_0_interpretedgenomescanc:[],org_gel_models_report_avro_4_0_0_interpretedgenomesrd:[],org_gel_models_report_avro_4_0_0_mdtdeliveryprotocol:[],org_gel_models_report_avro_4_0_0_rdparticip:[],org_gel_models_report_avro_4_0_0_rdparticipantchangelog:[],org_gel_models_report_avro_4_0_0_reportversioncontrol:[],org_gel_models_report_avro_4_0_0_supplementaryanalysisresult:[],org_gel_models_report_avro_4_1_0_auditlog:[],org_gel_models_report_avro_4_1_0_clinicalreportcanc:[],org_gel_models_report_avro_4_1_0_clinicalreportrd:[],org_gel_models_report_avro_4_1_0_commoninterpret:[],org_gel_models_report_avro_4_1_0_commoninterpretedcanc:[],org_gel_models_report_avro_4_1_0_commonrequest:[],org_gel_models_report_avro_4_1_0_exitquestionnair:[],org_gel_models_report_avro_4_1_0_interpretationrequestcanc:[],org_gel_models_report_avro_4_1_0_interpretationrequestrd:[],org_gel_models_report_avro_4_1_0_interpretedgenomescanc:[],org_gel_models_report_avro_4_1_0_interpretedgenomesrd:[],org_gel_models_report_avro_4_1_0_mdtdeliveryprotocol:[],org_gel_models_report_avro_4_1_0_reportversioncontrol:[],org_gel_models_report_avro_4_2_0:[],org_opencb_biodata_models_1_2_0_evid:[],org_opencb_biodata_models_1_2_0_read:[],org_opencb_biodata_models_1_2_0_vari:[],org_opencb_biodata_models_1_2_0_variantannot:[],organis:0,orggelmodelsreportavro210clinicalreportcanc:[],orggelmodelsreportavro210clinicalreportrd:[],other:[],our:0,out:0,pack:0,page:0,paramet:2,particip:0,participant_1_0_0:0,participantsensitiveinform:1,path_to:2,pedigre:[],pip:[0,2],plugin:0,posit:0,prepar:0,present:0,program:0,project:0,protocol:0,push:0,pyc:0,python:0,qual:0,rang:0,rare:[],rarediseaseinterpretationpipelin:1,raw:[],rdparticip:1,rdparticipantchangelog:1,read:1,readmethod:1,recent:0,record:0,refer:1,referencebas:0,referencemethod:1,referencenam:0,repo:0,report:0,reportedvari:[],reportev:1,reports_2_1_0:0,reports_3_0_0:0,reports_4_2_0:[],reports_5_0_0:0,reportversioncontrol:1,repositori:0,repres:0,request:0,rest:0,root:0,run:0,run_test:0,sampl:[],samplest:1,sandbox:0,schema:[],schema_model:2,script:[],see:0,self:0,separ:0,serial:[],serialis:0,servic:0,set:2,should:2,show:0,skip:0,snapshot:0,snapshot_auditlog:[],snapshot_cancerparticip:[],snapshot_clinicalreportcanc:[],snapshot_clinicalreportrd:[],snapshot_com:[],snapshot_common:[],snapshot_commoninterpret:[],snapshot_commoninterpretedcanc:[],snapshot_commonparticip:[],snapshot_commonrequest:[],snapshot_dataintakecanc:[],snapshot_dataintakerd:[],snapshot_evidenceset:[],snapshot_exitquestionnair:[],snapshot_gelbammetr:[],snapshot_gelvcfmetr:[],snapshot_individualst:[],snapshot_interpretationrequestcanc:[],snapshot_interpretationrequestrd:[],snapshot_interpretedgenomescanc:[],snapshot_interpretedgenomesrd:[],snapshot_mdtdeliveryprotocol:[],snapshot_metadata:[],snapshot_method:[],snapshot_observedvari:[],snapshot_participantsensitiveinform:[],snapshot_rarediseaseinterpretationpipelin:[],snapshot_rdparticip:[],snapshot_rdparticipantchangelog:[],snapshot_read:[],snapshot_readmethod:[],snapshot_refer:[],snapshot_referencemethod:[],snapshot_reportedvari:[],snapshot_reportversioncontrol:[],snapshot_samplest:[],snapshot_supplementaryanalysisresult:[],snapshot_transact:[],snapshot_vari:[],snapshot_variantmethod:[],snapshot_versioncontrol:[],some:0,sourc:0,specif:0,sphinx:0,sphynx:0,src:0,standard:0,statu:0,str:0,subcommand:0,success:0,sudo:[0,2],supplementaryanalysisresult:1,support:0,system:0,tab:0,target:0,technolog:0,them:0,thi:0,those:0,through:0,time:0,todo:0,tojsondict:0,tomcat:0,toolbox:0,total:0,transact:1,two:0,txt:0,under:0,updat:0,upgrad:2,usag:0,use:0,used:0,using:0,util:0,valid:[],valu:0,variant:0,variantannot:1,variantlegaci:1,variantmetadata:1,variantmethod:1,version:2,version_500:[],version_model:2,versioncontrol:1,want:2,war:0,which:0,window:0,yes:0,you:[0,2],your:0},titles:["GEL models documentation","GEL Models documentation","Models Validation"],titleterms:{"function":0,Using:0,addit:0,archiv:[],audit:[],avro:1,biodata:1,build:0,clinic:[],contain:0,coverag:1,custom:0,cva:1,data:0,document:[0,1],from:0,ga4gh:1,gel:[0,1],gel_bioinf_model:1,gelreportmodel:[],gener:[],genom:[],get:0,how:2,instal:[0,2],interpret:[],introduct:0,java:0,legaci:0,librari:2,main:0,metric:1,migrat:0,mock:0,model:[0,1,2],opencb:1,option:0,org:1,other:0,packag:0,particip:1,python:2,report:1,request:[],requir:0,resourc:0,schema:0,script:2,snapshot:1,start:0,system:1,test:0,tool:0,unit:0,use:2,valid:[0,2],variant:1,version:0}})
+Search.setIndex({docnames:["index","models","validation"],envversion:52,filenames:["index.rst","models.rst","validation.rst"],objects:{},objnames:{},objtypes:{},terms:{"4dabae77118d":0,"case":[],"class":0,"default":0,"import":0,"super":0,"true":0,"while":0,The:0,Then:0,There:0,Use:0,__init__:0,_blank:[],_metadata:1,_version:0,abov:0,access:0,actual:0,affect:0,against:0,agnost:0,aim:0,align:0,all:0,allow:0,alq:0,also:0,alternatebas:0,analysisreturnuri:0,ani:0,annot:0,antrun:0,applic:0,apt:0,archiv:0,arg:0,argument:0,assembl:0,asserttru:[],audit:[],auditlog:1,aug:0,automat:0,avail:0,avoid:0,avpr2html:0,avro:[],avro_factori:0,avrodoc:0,base:0,batch:0,between:0,biodata:0,box:0,build_model:0,busi:0,can:0,cancer:[],cancerinterpretationrequest:0,cancerinterpretationrequest_4_0_0:0,cancerinterpretationrequest_5_0_0:0,cancerinterpretedgenom:0,cancerparticip:1,catalog_variable_set:0,chang:[],check:0,cip:0,cipapi:0,clinic:0,clinicalreportcanc:1,clinicalreportrd:1,clone:0,code:0,com:0,command:0,comment:1,common:1,commoninterpret:1,commoninterpretedcanc:1,commonparticip:1,commonrequest:1,complet:0,configur:0,containerid:0,convers:0,conversion_tool:0,copi:0,correspond:0,could:0,coverag:0,creat:0,create_batch:0,cva:0,cva_0_3_0:0,data:[],dataintak:[],dataintakecanc:1,dataintakerd:1,def:0,demograph:0,depend:0,dependency_manag:0,deploi:0,describ:0,deserialis:0,dictionari:0,differ:0,directori:0,diseas:[],django:0,doc:0,docker:0,drwxr:0,e444d27c16b9:0,each:0,either:0,enabl:0,england:0,ensur:0,errr:0,evid:[0,1],evidenceset:1,exampl:0,exchang:0,exit:0,exitquestionnair:1,exitquestionnairebackup:[],extern:0,facilit:0,factori:0,factoryavro:0,field:0,file:0,fill:0,fill_nul:0,flag:0,folder:0,follow:0,format:0,fromjsondict:0,fromjsonstr:0,fuzz:0,fuzzi:0,fuzzychoic:0,fuzzyinteg:0,ga4gh:0,ga4gh_3_0_0:0,ga4ghvariantfactori:0,gel:[],gel_bioinf_model:0,gel_models_tool:[],gelbammetr:1,gelmodelstool:[],gelprotocol:0,gelreportmodel:[0,2],gelvcfmetr:1,gener:0,genericfactoryavro:0,genom:0,genomicsengland:0,get_factory_avro:0,git:0,github:0,given:0,grch38:0,greatli:0,handi:0,has:0,have:0,healthcheck:1,help:0,hoc:0,hold:0,how:[],href:[],html:0,html_schema:0,http:0,idl2avpr:0,idl2json:0,idl2python:0,idl:0,illustr:0,implement:0,includ:0,independ:0,index:0,individualst:1,inform:0,instal:[],instanc:0,instance_with_uri:0,intak:[],interfac:0,intern:0,interpret:0,interpretation_request_factori:[],interpretationdatard:[],interpretationrequestcanc:1,interpretationrequestrd:[0,1],interpretedgenomescanc:1,interpretedgenomesrd:1,ish:0,isvalid:0,its:0,jar:0,json2java:0,json2python:0,json:[],json_dict:0,json_fil:2,json_str:0,kwarg:0,lambda:0,land:0,languag:0,last:[0,2],latest:[0,2],lazyattribut:0,librari:0,list:0,log:[],mai:0,map:0,marker:0,maven:0,mdtdeliveryprotocol:1,memori:0,messag:0,meta:0,metadata:1,method:1,metric:0,metrics_1_0_0:0,migrate_cancer_interpretation_request:0,migrated_inst:0,migratereports400to500:0,migration_reports_4_0_0_to_reports_5_0_0:0,modelvalidatior:2,mvn:0,myuri:0,need:0,nodej:0,non:0,note:0,now:0,npm:0,nullabl:0,number:0,object:0,observ:[],observedvari:1,old_inst:0,onc:0,one:0,onli:0,opencb:0,opencb_1_2_0:0,org:0,org_ga4gh_models_3_0_0_common:[],org_ga4gh_models_3_0_0_metadata:[],org_ga4gh_models_3_0_0_method:[],org_ga4gh_models_3_0_0_read:[],org_ga4gh_models_3_0_0_readmethod:[],org_ga4gh_models_3_0_0_refer:[],org_ga4gh_models_3_0_0_referencemethod:[],org_ga4gh_models_3_0_0_vari:[],org_ga4gh_models_3_0_0_variantmethod:[],org_ga4gh_models_3_1_0:[],org_gel_models_cva_avro_0_3_1_com:[],org_gel_models_cva_avro_0_3_1_dataintak:[],org_gel_models_cva_avro_0_3_1_evidenceset:[],org_gel_models_cva_avro_0_3_1_observedvari:[],org_gel_models_cva_avro_0_3_1_reportedvari:[],org_gel_models_cva_avro_0_4_0:[],org_gel_models_metrics_avro_1_0_0_gelbammetr:[],org_gel_models_metrics_avro_1_0_0_gelvcfmetr:[],org_gel_models_metrics_avro_1_0_0_individualst:[],org_gel_models_metrics_avro_1_0_0_rarediseaseinterpretationpipelin:[],org_gel_models_metrics_avro_1_0_0_samplest:[],org_gel_models_metrics_avro_1_0_0_supplementaryanalysisresult:[],org_gel_models_metrics_avro_1_0_1_gelbammetr:[],org_gel_models_metrics_avro_1_0_1_gelvcfmetr:[],org_gel_models_metrics_avro_1_0_1_individualst:[],org_gel_models_metrics_avro_1_0_1_rarediseaseinterpretationpipelin:[],org_gel_models_metrics_avro_1_0_1_samplest:[],org_gel_models_metrics_avro_1_0_1_supplementaryanalysisresult:[],org_gel_models_metrics_avro_1_1_0:[],org_gel_models_participant_avro_1_0_0_cancerparticip:[],org_gel_models_participant_avro_1_0_0_commonparticip:[],org_gel_models_participant_avro_1_0_0_participantsensitiveinform:[],org_gel_models_participant_avro_1_0_0_rdparticip:[],org_gel_models_participant_avro_1_0_0_rdparticipantchangelog:[],org_gel_models_participant_avro_1_0_0_versioncontrol:[],org_gel_models_participant_avro_1_0_1_cancerparticip:[],org_gel_models_participant_avro_1_0_1_commonparticip:[],org_gel_models_participant_avro_1_0_1_participantsensitiveinform:[],org_gel_models_participant_avro_1_0_1_rdparticip:[],org_gel_models_participant_avro_1_0_1_rdparticipantchangelog:[],org_gel_models_participant_avro_1_0_1_versioncontrol:[],org_gel_models_participant_avro_1_0_2_cancerparticip:[],org_gel_models_participant_avro_1_0_2_commonparticip:[],org_gel_models_participant_avro_1_0_2_participantsensitiveinform:[],org_gel_models_participant_avro_1_0_2_rdparticip:[],org_gel_models_participant_avro_1_0_2_rdparticipantchangelog:[],org_gel_models_participant_avro_1_0_2_versioncontrol:[],org_gel_models_participant_avro_1_0_3:[],org_gel_models_report_avro_2_1_0_auditlog:[],org_gel_models_report_avro_2_1_0_cancerparticip:[],org_gel_models_report_avro_2_1_0_clinicalreportcanc:[],org_gel_models_report_avro_2_1_0_clinicalreportrd:[],org_gel_models_report_avro_2_1_0_commoninterpret:[],org_gel_models_report_avro_2_1_0_commoninterpretedcanc:[],org_gel_models_report_avro_2_1_0_commonparticip:[],org_gel_models_report_avro_2_1_0_commonrequest:[],org_gel_models_report_avro_2_1_0_exitquestionnair:[],org_gel_models_report_avro_2_1_0_gelbammetr:[],org_gel_models_report_avro_2_1_0_gelvcfmetr:[],org_gel_models_report_avro_2_1_0_individualst:[],org_gel_models_report_avro_2_1_0_interpretationrequestcanc:[],org_gel_models_report_avro_2_1_0_interpretationrequestrd:[],org_gel_models_report_avro_2_1_0_interpretedgenomescanc:[],org_gel_models_report_avro_2_1_0_interpretedgenomesrd:[],org_gel_models_report_avro_2_1_0_mdtdeliveryprotocol:[],org_gel_models_report_avro_2_1_0_participantsensitiveinform:[],org_gel_models_report_avro_2_1_0_rdparticip:[],org_gel_models_report_avro_2_1_0_rdparticipantchangelog:[],org_gel_models_report_avro_2_1_0_samplest:[],org_gel_models_report_avro_2_1_0_supplementaryanalysisresult:[],org_gel_models_report_avro_2_1_0_versioncontrol:[],org_gel_models_report_avro_3_0_0_auditlog:[],org_gel_models_report_avro_3_0_0_cancerparticip:[],org_gel_models_report_avro_3_0_0_clinicalreportcanc:[],org_gel_models_report_avro_3_0_0_clinicalreportrd:[],org_gel_models_report_avro_3_0_0_commoninterpret:[],org_gel_models_report_avro_3_0_0_commoninterpretedcanc:[],org_gel_models_report_avro_3_0_0_commonparticip:[],org_gel_models_report_avro_3_0_0_commonrequest:[],org_gel_models_report_avro_3_0_0_exitquestionnair:[],org_gel_models_report_avro_3_0_0_gelbammetr:[],org_gel_models_report_avro_3_0_0_gelvcfmetr:[],org_gel_models_report_avro_3_0_0_individualst:[],org_gel_models_report_avro_3_0_0_interpretationrequestcanc:[],org_gel_models_report_avro_3_0_0_interpretationrequestrd:[],org_gel_models_report_avro_3_0_0_interpretedgenomescanc:[],org_gel_models_report_avro_3_0_0_interpretedgenomesrd:[],org_gel_models_report_avro_3_0_0_mdtdeliveryprotocol:[],org_gel_models_report_avro_3_0_0_participantsensitiveinform:[],org_gel_models_report_avro_3_0_0_rdparticip:[],org_gel_models_report_avro_3_0_0_rdparticipantchangelog:[],org_gel_models_report_avro_3_0_0_samplest:[],org_gel_models_report_avro_3_0_0_supplementaryanalysisresult:[],org_gel_models_report_avro_3_0_0_versioncontrol:[],org_gel_models_report_avro_3_1_0_auditlog:[],org_gel_models_report_avro_3_1_0_cancerparticip:[],org_gel_models_report_avro_3_1_0_clinicalreportcanc:[],org_gel_models_report_avro_3_1_0_clinicalreportrd:[],org_gel_models_report_avro_3_1_0_commoninterpret:[],org_gel_models_report_avro_3_1_0_commoninterpretedcanc:[],org_gel_models_report_avro_3_1_0_commonparticip:[],org_gel_models_report_avro_3_1_0_commonrequest:[],org_gel_models_report_avro_3_1_0_exitquestionnair:[],org_gel_models_report_avro_3_1_0_gelbammetr:[],org_gel_models_report_avro_3_1_0_gelvcfmetr:[],org_gel_models_report_avro_3_1_0_individualst:[],org_gel_models_report_avro_3_1_0_interpretationrequestcanc:[],org_gel_models_report_avro_3_1_0_interpretationrequestrd:[],org_gel_models_report_avro_3_1_0_interpretedgenomescanc:[],org_gel_models_report_avro_3_1_0_interpretedgenomesrd:[],org_gel_models_report_avro_3_1_0_mdtdeliveryprotocol:[],org_gel_models_report_avro_3_1_0_participantsensitiveinform:[],org_gel_models_report_avro_3_1_0_rdparticip:[],org_gel_models_report_avro_3_1_0_rdparticipantchangelog:[],org_gel_models_report_avro_3_1_0_samplest:[],org_gel_models_report_avro_3_1_0_supplementaryanalysisresult:[],org_gel_models_report_avro_3_1_0_versioncontrol:[],org_gel_models_report_avro_4_0_0_auditlog:[],org_gel_models_report_avro_4_0_0_cancerparticip:[],org_gel_models_report_avro_4_0_0_clinicalreportcanc:[],org_gel_models_report_avro_4_0_0_clinicalreportrd:[],org_gel_models_report_avro_4_0_0_commoninterpret:[],org_gel_models_report_avro_4_0_0_commoninterpretedcanc:[],org_gel_models_report_avro_4_0_0_commonrequest:[],org_gel_models_report_avro_4_0_0_exitquestionnair:[],org_gel_models_report_avro_4_0_0_gelbammetr:[],org_gel_models_report_avro_4_0_0_interpretationrequestcanc:[],org_gel_models_report_avro_4_0_0_interpretationrequestrd:[],org_gel_models_report_avro_4_0_0_interpretedgenomescanc:[],org_gel_models_report_avro_4_0_0_interpretedgenomesrd:[],org_gel_models_report_avro_4_0_0_mdtdeliveryprotocol:[],org_gel_models_report_avro_4_0_0_rdparticip:[],org_gel_models_report_avro_4_0_0_rdparticipantchangelog:[],org_gel_models_report_avro_4_0_0_reportversioncontrol:[],org_gel_models_report_avro_4_0_0_supplementaryanalysisresult:[],org_gel_models_report_avro_4_1_0_auditlog:[],org_gel_models_report_avro_4_1_0_clinicalreportcanc:[],org_gel_models_report_avro_4_1_0_clinicalreportrd:[],org_gel_models_report_avro_4_1_0_commoninterpret:[],org_gel_models_report_avro_4_1_0_commoninterpretedcanc:[],org_gel_models_report_avro_4_1_0_commonrequest:[],org_gel_models_report_avro_4_1_0_exitquestionnair:[],org_gel_models_report_avro_4_1_0_interpretationrequestcanc:[],org_gel_models_report_avro_4_1_0_interpretationrequestrd:[],org_gel_models_report_avro_4_1_0_interpretedgenomescanc:[],org_gel_models_report_avro_4_1_0_interpretedgenomesrd:[],org_gel_models_report_avro_4_1_0_mdtdeliveryprotocol:[],org_gel_models_report_avro_4_1_0_reportversioncontrol:[],org_gel_models_report_avro_4_2_0:[],org_opencb_biodata_models_1_2_0_evid:[],org_opencb_biodata_models_1_2_0_read:[],org_opencb_biodata_models_1_2_0_vari:[],org_opencb_biodata_models_1_2_0_variantannot:[],organis:0,orggelmodelsreportavro210clinicalreportcanc:[],orggelmodelsreportavro210clinicalreportrd:[],other:[],our:0,out:0,pack:0,page:0,paramet:2,particip:0,participant_1_0_0:0,participantsensitiveinform:1,path_to:2,pedigre:[],pip:[0,2],pip_serv:0,plugin:0,posit:0,prepar:0,present:0,print:0,program:0,project:0,protocol:0,push:0,pyc:0,python:0,qual:0,rang:0,rare:[],rarediseaseinterpretationpipelin:1,raw:[],rdparticip:1,rdparticipantchangelog:1,read:1,readmethod:1,recent:0,record:0,refer:1,referencebas:0,referencemethod:1,referencenam:0,repo:0,report:0,reportedvari:[],reportev:1,reports_2_1_0:0,reports_3_0_0:0,reports_4_0_0:0,reports_4_2_0:[],reports_5_0_0:0,reportversioncontrol:1,repositori:0,repres:0,request:0,requir:0,rest:0,root:0,rule:0,run:0,run_test:0,sampl:[],samplest:1,sandbox:0,schema:[],schema_model:2,script:0,see:0,self:0,separ:0,serial:[],serialis:[],servic:0,set:2,should:2,show:0,skip:0,snapshot:0,snapshot_auditlog:[],snapshot_cancerparticip:[],snapshot_clinicalreportcanc:[],snapshot_clinicalreportrd:[],snapshot_com:[],snapshot_common:[],snapshot_commoninterpret:[],snapshot_commoninterpretedcanc:[],snapshot_commonparticip:[],snapshot_commonrequest:[],snapshot_dataintakecanc:[],snapshot_dataintakerd:[],snapshot_evidenceset:[],snapshot_exitquestionnair:[],snapshot_gelbammetr:[],snapshot_gelvcfmetr:[],snapshot_individualst:[],snapshot_interpretationrequestcanc:[],snapshot_interpretationrequestrd:[],snapshot_interpretedgenomescanc:[],snapshot_interpretedgenomesrd:[],snapshot_mdtdeliveryprotocol:[],snapshot_metadata:[],snapshot_method:[],snapshot_observedvari:[],snapshot_participantsensitiveinform:[],snapshot_rarediseaseinterpretationpipelin:[],snapshot_rdparticip:[],snapshot_rdparticipantchangelog:[],snapshot_read:[],snapshot_readmethod:[],snapshot_refer:[],snapshot_referencemethod:[],snapshot_reportedvari:[],snapshot_reportversioncontrol:[],snapshot_samplest:[],snapshot_supplementaryanalysisresult:[],snapshot_transact:[],snapshot_vari:[],snapshot_variantmethod:[],snapshot_versioncontrol:[],some:0,sourc:0,specif:0,sphinx:0,sphynx:0,src:0,standard:0,statu:0,str:0,string:0,subcommand:0,success:0,sudo:[0,2],supplementaryanalysisresult:1,support:0,system:0,tab:0,target:0,technolog:0,thei:0,them:0,thi:0,those:0,through:0,tieredvari:0,time:0,todo:0,tojsondict:0,tojsonstr:0,tomcat:0,toolbox:0,total:0,transact:1,two:0,txt:0,under:0,updat:0,update_docs_index:0,upgrad:2,usag:0,use:0,used:0,using:0,util:0,valid:[],valu:0,variant:0,variantannot:1,variantlegaci:1,variantmetadata:1,variantmethod:1,verbos:0,version:2,version_400:0,version_500:[],version_model:2,versioncontrol:1,want:2,war:0,which:0,window:0,yes:0,you:[0,2],your:0},titles:["GEL models documentation","GEL Models documentation","Models Validation"],titleterms:{"function":0,Using:0,addit:0,archiv:[],audit:[],avro:[0,1],biodata:1,build:0,built:0,clinic:[],contain:0,coverag:1,custom:0,cva:1,data:0,document:[0,1],from:0,ga4gh:1,gel:[0,1],gel_bioinf_model:1,gelreportmodel:[],gener:[],genom:[],get:0,how:2,instal:[0,2],interpret:[],introduct:0,java:0,json:0,legaci:0,librari:2,local:0,main:0,metric:1,migrat:0,mock:0,model:[0,1,2],opencb:1,option:0,org:1,other:0,packag:0,particip:1,pypip:0,python:2,report:1,request:[],requir:[],resourc:0,schema:0,script:2,serialis:0,snapshot:1,start:0,system:1,test:0,tool:0,unit:0,use:2,valid:[0,2],variant:1,version:0}})
