+++ setup.py
@@ -20,7 +20,7 @@ setup(
     version=VERSION,
     packages=find_packages(),
     include_package_data=True,
-    scripts=['bin/ModelValidator'],
+    scripts=['bin/ModelValidator', 'protocols_utils/utils/conversion_tools.py'],
     url='',
     license='',
     author='Bioinformatics Team at Genomics England',
